move globals to globals.c - minor cleanup in clifm.c - minor changes to helpers.h
